refactor:Moving form to its own component
